[SHIBBOLETH] Registering basic math types with script.
NOTE: If math perf ever gets bad in scripts, will want to profile to see if calling into C++ for math operations is too costly.

+++ src/Entity/Shibboleth_Object.cpp
@@ -265,9 +265,9 @@ void Object::setLocalAABB(const Gleam::AABB& aabb)
 	markDirty();
 }
 
-size_t Object::getNumComponents(void) const
+int32_t Object::getNumComponents(void) const
 {
-	return _components.size();
+	return static_cast<int32_t>(_components.size());
 }
 
 const Component* Object::getComponent(int32_t index) const
